+++ tests/dfsvisitortest.c
@@ -54,7 +54,7 @@ test_dfs_visitor_traverse(Test *test)
     visits_list = gegl_visitor_get_visits_list(GEGL_VISITOR(mock_dfs_visitor));
 
     ct_test (test, 7 == g_list_length(visits_list));
-    for(i = 0; i < g_list_length(visits_list); i++)
+    for(i = 0; i < (gint)g_list_length(visits_list); i++)
       {
         GeglNode *node = (GeglNode*)g_list_nth_data(visits_list, i);
         const gchar *name = gegl_object_get_name(GEGL_OBJECT(node));
@@ -74,7 +74,7 @@ test_dfs_visitor_traverse(Test *test)
     visits_list = gegl_visitor_get_visits_list(GEGL_VISITOR(mock_dfs_visitor));
 
     ct_test (test, 4 == g_list_length(visits_list));
-    for(i = 0; i < g_list_length(visits_list); i++)
+    for(i = 0; i < (gint)g_list_length(visits_list); i++)
       {
         GeglNode *node = (GeglNode*)g_list_nth_data(visits_list, i);
         const gchar *name = gegl_object_get_name(GEGL_OBJECT(node));
@@ -115,7 +115,7 @@ test_dfs_visitor_traverse_graph(Test *test)
     visits_list = gegl_visitor_get_visits_list(GEGL_VISITOR(mock_dfs_visitor));
 
     ct_test (test, 7 == g_list_length(visits_list));
-    for(i = 0; i < g_list_length(visits_list); i++)
+    for(i = 0; i < (gint)g_list_length(visits_list); i++)
       {
         GeglNode *node = (GeglNode*)g_list_nth_data(visits_list, i);
         const gchar *name = gegl_object_get_name(GEGL_OBJECT(node));
@@ -151,7 +151,7 @@ test_dfs_visitor_traverse_diamond(Test *test)
     visits_list = gegl_visitor_get_visits_list(GEGL_VISITOR(mock_dfs_visitor));
 
     ct_test (test, 5 == g_list_length(visits_list));
-    for(i = 0; i < g_list_length(visits_list); i++)
+    for(i = 0; i < (gint)g_list_length(visits_list); i++)
       {
         GeglNode *node = (GeglNode*)g_list_nth_data(visits_list, i);
         const gchar *name = gegl_object_get_name(GEGL_OBJECT(node));
@@ -213,7 +213,7 @@ test_dfs_visitor_simple_shared(Test *test)
     visits_list = gegl_visitor_get_visits_list(GEGL_VISITOR(mock_dfs_visitor));
 
     ct_test (test, 4 == g_list_length(visits_list));
-    for(i = 0; i < g_list_length(visits_list); i++)
+    for(i = 0; i < (gint)g_list_length(visits_list); i++)
       {
         GeglNode *node = (GeglNode*)g_list_nth_data(visits_list, i);
         const gchar *name = gegl_object_get_name(GEGL_OBJECT(node));
@@ -254,7 +254,7 @@ test_dfs_visitor_traverse_with_nulls(Test *test)
     visits_list = gegl_visitor_get_visits_list(GEGL_VISITOR(mock_dfs_visitor));
 
     ct_test (test, 4 == g_list_length(visits_list));
-    for(i = 0; i < g_list_length(visits_list); i++)
+    for(i = 0; i < (gint)g_list_length(visits_list); i++)
       {
         GeglNode *node = (GeglNode*)g_list_nth_data(visits_list, i);
         const gchar *name = gegl_object_get_name(GEGL_OBJECT(node));
